+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.5] - 2022-05-25
+
 ### Added
 
 - Added One Minus float extension to do one minus (1-x) operations.
+- Added Random List extension that takes excluded items as a parameter.
 
 ## [0.1.4] - 2022-05-03